requirements.txt
google-generativeai>=0.7.2
duckduckgo-search>=6.2.6
python-dotenv>=1.0.1
